Extract ListStyleModifier to avod duplicated code
Refactor ListStyleModifier into a separate file for improved organization and reusability across DynamicList and SectionedDynamicList components. Removed redundant implementations from both content files to streamline code and enhance maintainability.
