+++ Sources/DynamicList/Private/UI/Dynamic List/DynamicListContent.swift
@@ -152,26 +152,4 @@ struct DynamicListContent<Item: Identifiable & Hashable>: View {
             DefaultErrorView(error: error)
         }
     }
-
-    /// ViewModifier to apply list styles
-    private struct ListStyleModifier: ViewModifier {
-        let style: ListStyleType
-
-        func body(content: Content) -> some View {
-            switch style {
-            case .automatic:
-                content.listStyle(.automatic)
-            case .plain:
-                content.listStyle(.plain)
-            case .inset:
-                content.listStyle(.inset)
-            #if os(iOS)
-            case .grouped:
-                content.listStyle(.grouped)
-            case .insetGrouped:
-                content.listStyle(.insetGrouped)
-            #endif
-            }
-        }
-    }
 }
